+++ posthog-android/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: allow override of event timestamp ([#309](https://github.com/PostHog/posthog-android/pull/309))
+
 ## 3.24.0 - 2025-10-20
 
 - add evaluation tags to android SDK ([#301](https://github.com/PostHog/posthog-android/pull/301))